+++ base-crypto/src/schnorr.rs
@@ -13,11 +13,11 @@
 
 //! Schnorr signature scheme over secp256k1, conforming to BIP340.
 use crate::BinaryHashRepr;
+use crate::rng::derive_crypto_rng;
 use k256::schnorr;
 #[cfg(feature = "proptest")]
 use proptest::arbitrary::Arbitrary;
 use rand::distributions::{Distribution, Standard};
-use rand::rngs::OsRng;
 use rand::{CryptoRng, Rng};
 use serde::{Deserialize, Serialize};
 use serialize::{Deserializable, Serializable, Tagged, VecExt, tag_enforcement_test};
@@ -96,8 +96,8 @@ simple_arbitrary!(VerifyingKey);
 serialize::randomised_serialization_test!(VerifyingKey);
 
 impl Distribution<VerifyingKey> for Standard {
-    fn sample<R: Rng + ?Sized>(&self, _rng: &mut R) -> VerifyingKey {
-        SigningKey::sample(OsRng).verifying_key()
+    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> VerifyingKey {
+        SigningKey::sample(derive_crypto_rng(rng)).verifying_key()
     }
 }
 
@@ -242,10 +242,11 @@ serialize::randomised_serialization_test!(Signature);
 
 impl Distribution<Signature> for Standard {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signature {
-        let signing_key = SigningKey::sample(OsRng);
+        let mut crypto_rng = derive_crypto_rng(rng);
+        let signing_key = SigningKey::sample(&mut crypto_rng);
         let mut message = Vec::with_bounded_capacity(32);
         rng.fill_bytes(&mut message);
-        signing_key.sign(&mut OsRng, &message)
+        signing_key.sign(&mut crypto_rng, &message)
     }
 }